Refactor components
Added components folder
Moved App.js to components folder
Added SelectOption file and component into components folder
