+++ src/components/App.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import './App.css';
+import '../App.css';
+import SelectOption from './SelectOption'
 
 
 const axios = require('axios')
@@ -146,28 +147,7 @@ class App extends React.Component {
   }
 }
 
-function SelectOption(props) {
-  return (
-    <div>
-      <h2>Find a random anime or manga:</h2>
-      <select id="genreOption">
-        <option value=""></option>
-        <option value="anime">Anime</option>
-        <option value="manga">Manga</option>
-      </select>
-      <br />
-      <button onClick={props.gRequest}>Get Request</button>
-      <p id="errorDesc">{props.errorDescr}</p>
-      <section id="showResults"></section>
-      <div>
-        <br />
-        <img src={props.data[1]}></img>
-        <h3>{props.data[0]}</h3>
-        <p id="searchSummary">{props.data[2]}</p>
-      </div>
-    </div>
-  )
-}
+
 
 /*
 function SearchOption(){
